requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
aioredis==2.0.1
httpx==0.25.1
cryptography==41.0.5
python-multipart==0.0.6
pymilvus==2.3.3
alembic==1.12.1
aio-pika==9.3.0
apscheduler==3.10.4
prometheus-client==0.19.0
sqlalchemy[asyncio]==2.0.23
psutil==5.9.6
pandas==2.1.3
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosmtplib==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0